refactor(sync-devices): use local PlatformExampleDeviceEntry type
