+++ tools/scripts/sync-devices/src/types.ts
@@ -1,0 +1,7 @@
+import type { ExampleInfo } from '@chirimen-device-dashboard/shared-types';
+
+export interface PlatformExampleDeviceEntry {
+  dashboardDeviceId: string;
+  exampleDeviceId: string;
+  examples: ExampleInfo[];
+}
